Se agrega módulo de novedades.
