+++ app/src/main/AndroidManifest.xml
@@ -2,14 +2,14 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="utn.frba.mobile.konzern">
 
-    <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE"/>
+    <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />
     <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" />
 
     <application
-        android:requestLegacyExternalStorage="true"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
+        android:requestLegacyExternalStorage="true"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
@@ -21,24 +21,29 @@
             android:grantUriPermissions="true">
             <meta-data
                 android:name="android.support.FILE_PROVIDER_PATHS"
-                android:resource="@xml/file_paths"></meta-data>
+                android:resource="@xml/file_paths" />
         </provider>
 
-        <activity android:name=".splash.SplashActivity"
+        <activity
+            android:name=".splash.SplashActivity"
             android:theme="@style/Splash">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
+
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".MainActivity"/>
-        <activity android:name=".login.LoginActivity"/>
+        <activity android:name=".MainActivity" />
+        <activity android:name=".login.LoginActivity" />
         <activity android:name=".profile.ProfileActivity" />
-        <activity android:name=".posts.PostActivity"
-            android:windowSoftInputMode="adjustPan"/>
-        <activity android:name=".reservations.ReservationsActivity"
-            android:windowSoftInputMode="adjustPan"/>
-        <activity android:name=".expenses.ExpensesActivity"/>
-        <activity android:name=".contact.ContactActivity"/>
+        <activity android:name=".posts.PostActivity"/>
+        <activity
+            android:name=".reservations.ReservationsActivity"
+            android:windowSoftInputMode="adjustPan" />
+        <activity android:name=".expenses.ExpensesActivity" />
+        <activity android:name=".contact.ContactActivity" />
+        <activity android:name=".posts.BaseActivity"/>
+        <activity android:name=".news.NewsActivity"/>
     </application>
+
 </manifest>
