+++ app/src/main/java/utn/frba/mobile/konzern/home/HomeFragment.kt
@@ -14,13 +14,16 @@ import kotlinx.android.synthetic.main.fragment_home.*
 import utn.frba.mobile.konzern.R
 import utn.frba.mobile.konzern.home.adapter.HomePostItemAdapter
 import utn.frba.mobile.konzern.home.adapter.LoadingItemAdapter
+import utn.frba.mobile.konzern.news.NewsActivity
 import utn.frba.mobile.konzern.posts.PostActivity
 import utn.frba.mobile.konzern.posts.model.Post
+import utn.frba.mobile.konzern.news.NewsViewModel
 import utn.frba.mobile.konzern.posts.viewModel.PostViewModel
 import utn.frba.mobile.konzern.utils.ScaleLayoutManager
 
 class HomeFragment : Fragment(), HomePostItemAdapter.OnHomeItemPostClickListener {
     private val postViewModel: PostViewModel by activityViewModels()
+    private val newsViewModel: NewsViewModel by activityViewModels()
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
         return inflater.inflate(R.layout.fragment_home, container, false)
@@ -29,6 +32,11 @@ class HomeFragment : Fragment(), HomePostItemAdapter.OnHomeItemPostClickListener
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
+        setPostModuleView()
+        setNewsModuleView()
+    }
+
+    private fun setPostModuleView(){
         postViewModel.itemList.observe(viewLifecycleOwner, Observer<List<Post>> {
             setPostRecycler(it)
         })
@@ -40,10 +48,20 @@ class HomeFragment : Fragment(), HomePostItemAdapter.OnHomeItemPostClickListener
         })
 
         postViewModel.initItemList()
-        setButtonsListeners()
     }
 
-    private fun setButtonsListeners(){
+    private fun setNewsModuleView(){
+        newsViewModel.itemList.observe(viewLifecycleOwner, Observer<List<Post>> {
+            setNewsRecycler(it)
+        })
+
+        newsViewModel.showProgressIndicator.observe(viewLifecycleOwner, Observer {
+            if(it!!){
+                vRecyclerViewNewsHome.adapter = LoadingItemAdapter()
+            }
+        })
+
+        newsViewModel.initItemList()
     }
 
     override fun onItemClick(id: String) {
@@ -51,16 +69,29 @@ class HomeFragment : Fragment(), HomePostItemAdapter.OnHomeItemPostClickListener
         findNavController().navigate(R.id.action_HomeFragment_to_ItemPostFragment)
     }
 
-    private fun setPostRecycler(items: List<Post>){
-        vRecyclerViewPostHome.adapter = HomePostItemAdapter(items, this)
-        vRecyclerViewPostHome.layoutManager = ScaleLayoutManager(requireActivity(), RecyclerView.HORIZONTAL, false, 0.15f, 0.85f)
+    private fun setRecycler(items: List<Post>, recyclerView: RecyclerView){
+        recyclerView.adapter = HomePostItemAdapter(items, this)
+        recyclerView.layoutManager = ScaleLayoutManager(requireActivity(), RecyclerView.HORIZONTAL, false, 0.15f, 0.85f)
 
         val snapHelperCardsRecycler = androidx.recyclerview.widget.PagerSnapHelper()
-        snapHelperCardsRecycler.attachToRecyclerView(vRecyclerViewPostHome)
-        this.vRecyclerViewPostHome.addItemDecoration(SnapItemDecorator(requireActivity()))
+        snapHelperCardsRecycler.attachToRecyclerView(recyclerView)
+        recyclerView.addItemDecoration(SnapItemDecorator(requireActivity()))
+    }
+
+    private fun setPostRecycler(items: List<Post>){
+        setRecycler(items, vRecyclerViewPostHome)
 
         vButtonPostMain.setOnClickListener{
             val intent = Intent(requireContext(), PostActivity::class.java)
+            this.startActivity(intent)
+        }
+    }
+
+    private fun setNewsRecycler(items: List<Post>){
+        setRecycler(items, vRecyclerViewNewsHome)
+
+        vButtonNewsMain.setOnClickListener{
+            val intent = Intent(requireContext(), NewsActivity::class.java)
             this.startActivity(intent)
         }
     }
