+++ app/src/main/java/utn/frba/mobile/konzern/posts/PostActivity.kt
@@ -1,67 +1,31 @@
 package utn.frba.mobile.konzern.posts
 
-import android.content.Intent
-import android.os.Bundle
-import android.view.View.GONE
-import android.view.View.VISIBLE
+import android.view.View
 import androidx.activity.viewModels
-import androidx.appcompat.app.AppCompatActivity
-import androidx.lifecycle.Observer
 import kotlinx.android.synthetic.main.activity_post.*
 import utn.frba.mobile.konzern.R
-import utn.frba.mobile.konzern.contact.ContactActivity
-import utn.frba.mobile.konzern.customviews.ToolbarMenuInterface
 import utn.frba.mobile.konzern.posts.viewModel.PostViewModel
-import utn.frba.mobile.konzern.profile.ProfileActivity
-import utn.frba.mobile.konzern.reservations.ReservationsActivity
+import utn.frba.mobile.konzern.utils.BaseViewModel
 
-class PostActivity : AppCompatActivity(), ToolbarMenuInterface {
-    val viewModel: PostViewModel by viewModels()
+class PostActivity : BaseActivity() {
+    override val title: String? = "Publicaciones"
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_post)
-
-        vActivityPostToolbar.setTitle(getString(R.string.posts_toolbar_title))
-
-        viewModel.showProgressIndicator.observe(this, Observer {
-            if(it!!){
-                showProgress()
-            }else{
-                hideProgress()
-            }
-        })
+    override fun getCustomViewModel(): BaseViewModel {
+        val vm: PostViewModel by viewModels()
+        return vm
     }
 
-    private fun showProgress(){
-        vActivityContentPost.visibility = GONE
-        vPostProgressBarLayout.visibility = VISIBLE
+    override fun getContentLayout(): Int {
+        return R.layout.activity_post
     }
 
-    private fun hideProgress(){
-        vPostProgressBarLayout.visibility = GONE
-        vActivityContentPost.visibility = VISIBLE
+    override fun showProgress(){
+        vActivityContentPost.visibility = View.GONE
+        vPostProgressBarLayout.visibility = View.VISIBLE
     }
 
-    override fun onMenuMyProfileClicked() {
-        val intent = Intent(this, ProfileActivity::class.java)
-        startActivity(intent)
-        finish()
-    }
-
-    override fun onMenuSeeReservationsClicked() {
-        val intent = Intent(this, ReservationsActivity::class.java)
-        startActivity(intent)
-        finish()
-    }
-
-    override fun onToolbarLogoClicked() {
-        finish()
-    }
-
-    override fun onMenuContactInfoClicked() {
-        val intent = Intent(this, ContactActivity::class.java)
-        startActivity(intent)
-        finish()
+    override fun hideProgress(){
+        vPostProgressBarLayout.visibility = View.GONE
+        vActivityContentPost.visibility = View.VISIBLE
     }
 }
