+++ app/src/main/java/utn/frba/mobile/konzern/posts/repository/NewsRepository.kt
@@ -1,0 +1,5 @@
+package utn.frba.mobile.konzern.posts.repository
+
+class NewsRepository: BasePostRepository() {
+    override var dbCollectionName = "news"
+}
