+++ app/src/main/java/utn/frba/mobile/konzern/posts/ui/PostBaseFragment.kt
@@ -1,9 +1,16 @@
 package utn.frba.mobile.konzern.posts.ui
 
+import android.os.Bundle
 import androidx.fragment.app.Fragment
-import androidx.fragment.app.activityViewModels
-import utn.frba.mobile.konzern.posts.viewModel.PostViewModel
+import utn.frba.mobile.konzern.posts.viewModel.BasePostViewModel
 
 abstract class PostBaseFragment: Fragment() {
-    protected val viewModel: PostViewModel by activityViewModels()
+    protected lateinit var viewModel: BasePostViewModel
+
+    abstract fun getCustomViewModel(): BasePostViewModel
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        viewModel = getCustomViewModel()
+    }
 }
