+++ app/src/main/java/utn/frba/mobile/konzern/posts/ui/PostDetailFragment.kt
@@ -6,14 +6,24 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.core.net.toUri
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.activityViewModels
 import androidx.lifecycle.Observer
 import androidx.lifecycle.observe
 import kotlinx.android.synthetic.main.fragment_post_detail.*
 import utn.frba.mobile.konzern.R
 import utn.frba.mobile.konzern.posts.adapter.ImageSliderAdapter
 import utn.frba.mobile.konzern.posts.model.Post
+import utn.frba.mobile.konzern.posts.viewModel.BasePostViewModel
+import utn.frba.mobile.konzern.posts.viewModel.PostViewModel
 
-class PostDetailFragment : PostBaseFragment() {
+open class PostDetailFragment : PostBaseFragment() {
+
+    override fun getCustomViewModel(): BasePostViewModel {
+        val vm: PostViewModel by activityViewModels()
+        return vm
+    }
+
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
         return inflater.inflate(R.layout.fragment_post_detail, container, false)
     }
